Move into modules and components folders
