+++ src/index.tsx
@@ -3,23 +3,23 @@ import AsyncCard from './components/AsyncCard';
 import TempAttachments from './modules/Attachments/TempAttachments';
 import Attachments from './modules/Attachments/Attachments';
 import CollapsibleCard from './components/CollapsibleCard';
-import InfoItem from './InfoItem/InfoItem';
-import StatusColumn from './InfoItem/StatusColumn';
+import InfoItem from './components/InfoItem/InfoItem';
+import StatusColumn from './components/InfoItem/StatusColumn';
 import ErrorPage from './components/error/ErrorPage';
 import HomeButton from './components/buttons/HomeButton';
 import ReloadButton from './components/buttons/ReloadButton';
 import BackButton from './components/buttons/BackButton';
 import ProcosysButton from './components/buttons/ProcosysButton';
-import Navbar from './NavBar/NavBar';
+import Navbar from './components/NavBar';
 import LoadingPage from './components/loading/LoadingPage';
 import SkeletonLoadingPage from './components/loading/SkeletonLoader';
 import removeSubdirectories from './utils/removeSubdirectories';
-import EntityDetails from './EntityDetails/EntityDetails';
-import TextIcon from './EntityDetails/TextIcon';
+import EntityDetails from './components/EntityDetails/EntityDetails';
+import TextIcon from './components/EntityDetails/TextIcon';
 import SearchTypeButton from './components/buttons/SearchTypeButton';
 import { StorageKey } from './typings/enums';
 import PageHeader from './components/PageHeader';
-import TagPhotoRecognition from './TagOcr/TagPhotoRecognition';
+import TagPhotoRecognition from './modules/TagOcr/TagPhotoRecognition';
 
 export {
     Checklist,
