+++ src/modules/TagOcr/TagPhotoRecognition.tsx
@@ -2,8 +2,8 @@ import { Button } from '@equinor/eds-core-react';
 import axios from 'axios';
 import React, { useRef, useState } from 'react';
 import styled from 'styled-components';
-import EdsIcon from '../components/icons/EdsIcon';
-import { AsyncStatus } from '../typings/enums';
+import EdsIcon from '../../components/icons/EdsIcon';
+import { AsyncStatus } from '../../typings/enums';
 import TagSelectionModal from './TagSelectionModal';
 
 const CaptureInput = styled.input`
